Expand PR template
Mostly I pulled in some things from https://github.com/flarum/org/issues/47. I didn't want to include all of that to reduce overhead.

One thing I'm not sure on is the merge process / review criteria documentation. My gut feeling is that these would be better in a docs page, and we could include a link in the actual template.