docs: increase logo resolution
README.md is using the logo with
the width of 400px, but the logo is only 298px,
so it looks blurry. Here, it's being replaced
by a higher resolution version.
I downloaded the .svg version of the logo from
https://github.com/flarum/framework/blob/main/framework/core/views/install/logo.svg
and then exported to png with the width of 400px.
However, 400px version didn't look
as crisp as it could, so I increased it to
around 1000px so that it looks perfectly crisp.

Signed-off-by: Soobin Rho <soobinrho@gmail.com>